requirements.txt
openai
pandas
openpyxl
python-dotenv
pytest
streamlit
python-docx
transformers
torch
